backend/requirements.txt
Django==5.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.6.0
drf-yasg==1.21.7
psycopg==3.2.3
Pillow==10.4.0
django-storages==1.14.2
minio==7.2.0
python-decouple==3.8